Relocate CollectionType out of registry to the root level
