+++ python/lsst/daf/butler/_collection_type.py
