+++ python/lsst/daf/butler/registry/collections/_base.py
@@ -37,10 +37,10 @@ from typing import TYPE_CHECKING, Any, Generic, Literal, NamedTuple, TypeVar, ca
 
 import sqlalchemy
 
+from ..._collection_type import CollectionType
 from ..._exceptions import CollectionCycleError, CollectionTypeError, MissingCollectionError
 from ...timespan_database_representation import TimespanDatabaseRepresentation
 from .._collection_record_cache import CollectionRecordCache
-from .._collection_type import CollectionType
 from ..interfaces import ChainedCollectionRecord, CollectionManager, CollectionRecord, RunRecord, VersionTuple
 from ..wildcards import CollectionWildcard
