+++ python/lsst/daf/butler/registry/collections/nameKey.py
@@ -35,9 +35,9 @@ from typing import TYPE_CHECKING, Any
 import sqlalchemy
 
 from ... import ddl
+from ..._collection_type import CollectionType
 from ...column_spec import COLLECTION_NAME_MAX_LENGTH
 from ...timespan_database_representation import TimespanDatabaseRepresentation
-from .._collection_type import CollectionType
 from ..interfaces import ChainedCollectionRecord, CollectionRecord, RunRecord, VersionTuple
 from ._base import (
     CollectionTablesTuple,
