+++ python/lsst/daf/butler/registry/queries/_sql_query_backend.py
@@ -34,13 +34,13 @@ from typing import TYPE_CHECKING, Any, cast
 
 from lsst.daf.relation import ColumnError, ColumnExpression, ColumnTag, Join, Predicate, Relation
 
+from ..._collection_type import CollectionType
 from ..._column_categorization import ColumnCategorization
 from ..._column_tags import DimensionKeyColumnTag, DimensionRecordColumnTag
 from ..._dataset_type import DatasetType
 from ..._exceptions import DataIdValueError
 from ...dimensions import DimensionGroup, DimensionRecordSet, DimensionUniverse
 from ...dimensions.record_cache import DimensionRecordCache
-from .._collection_type import CollectionType
 from ..interfaces import CollectionRecord, Database
 from ._query_backend import QueryBackend
 from ._sql_query_context import SqlQueryContext
