+++ python/lsst/daf/butler/script/removeRuns.py
@@ -32,7 +32,8 @@ from dataclasses import dataclass
 from functools import partial
 
 from .._butler import Butler
-from ..registry import CollectionType, MissingCollectionError
+from .._collection_type import CollectionType
+from ..registry import MissingCollectionError
 
 
 @dataclass
