+++ python/lsst/daf/butler/tests/butler_queries.py
@@ -41,6 +41,7 @@ import astropy.time
 from lsst.sphgeom import Region
 
 from .._butler import Butler
+from .._collection_type import CollectionType
 from .._dataset_type import DatasetType
 from .._exceptions import InvalidQueryError
 from .._timespan import Timespan
@@ -48,7 +49,7 @@ from ..dimensions import DataCoordinate, DimensionGroup, DimensionRecord
 from ..direct_query_driver import DirectQueryDriver
 from ..queries import DimensionRecordQueryResults
 from ..queries.tree import Predicate
-from ..registry import CollectionType, NoDefaultCollectionError, RegistryDefaults
+from ..registry import NoDefaultCollectionError, RegistryDefaults
 from ..registry.sql_registry import SqlRegistry
 from ..transfers import YamlRepoImportBackend
 from .utils import TestCaseMixin
