+++ python/lsst/daf/butler/transfers/_interfaces.py
@@ -33,13 +33,13 @@ from abc import ABC, abstractmethod
 from collections.abc import Iterable
 from typing import TYPE_CHECKING
 
+from .._collection_type import CollectionType
 from .._config import ConfigSubset
 from .._dataset_association import DatasetAssociation
 from .._dataset_type import DatasetType
 from .._file_dataset import FileDataset
 from ..datastore import Datastore
 from ..dimensions import DimensionElement, DimensionRecord
-from ..registry import CollectionType
 from ..registry.interfaces import CollectionRecord
 
 if TYPE_CHECKING:
